src/requirements.txt
aiofiles==24.1.0
aiogram==3.13.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
attrs==24.2.0
black==24.10.0
certifi==2024.8.30
click==8.1.7
frozenlist==1.5.0
idna==3.10
magic-filter==1.0.12
multidict==6.1.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
propcache==0.2.0
pyaes==1.6.1
pydantic==2.9.2
pydantic_core==2.23.4
PySocks==1.7.1
python-dotenv==1.0.1
TgCrypto==1.2.5
typing_extensions==4.12.2
yarl==1.16.0
